feat: add database migration files and configuration for drizzle-orm
This commit adds the necessary files for database migration using drizzle-orm. It includes the migration files for creating tables for items, shoppings, and users, as well as the necessary foreign key constraints. It also includes the configuration file for drizzle-orm, which specifies the output directory for the migration files and the schema directory. Additionally, a migration script is added to run the migrations using drizzle-orm. Finally, the .env file is added to the .gitignore file to prevent sensitive information from being committed to the repository.

chore: reformat tsconfig.json to improve readability and add exclusion of seed.ts from compilation